pyproject.toml
[project]
name = "arc-agi-3-solver"
version = "0.1.0"
description = "ARC-AGI-3 solver using depth-first search with nightshift-powered continuous improvement"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "arc-agi>=0.9.3",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]
kaggle = [
    "kaggle>=1.6.0",
]

[project.scripts]
arc-solve = "main:cli"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
